feat: simplify AdminDashboardContent by managing tab state within the component
